# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  Entity.Role:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: string
      name:
        type: string
      permissionsId:
        items:
          type: string
        type: array
      updatedAt:
        type: string
    type: object
...
  Response.GeneralResponse:
    properties:
      data: {}
      error:
        type: boolean
      message:
        type: string
    required:
    - data
    - error
    - message
    type: object
...
host: localhost:3000
info:
  contact:
    email: team@supnex.com
    name: Supnex
    url: https://supne.com/
  description: Golang Monolithic Boilerplate
  title: Golang Monolithic Boilerplate
  version: "1.0"
paths:
...
  /role-permission/attach/{roleId}{permissionId}:
    patch:
      consumes:
      - application/json
      description: Attach permission to role with roleId and permissionId
      parameters:
      - description: roleId
        in: path
        name: roleId
        required: true
        type: string
      - description: permissionId
        in: path
        name: permissionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/Entity.Role'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Attach permission to role
      tags:
      - Role-Permission
  /role-permission/detach/{roleId}{permissionId}:
    patch:
      consumes:
      - application/json
      description: Detach permission from role with roleId and permissionId
      parameters:
      - description: roleId
        in: path
        name: roleId
        required: true
        type: string
      - description: permissionId
        in: path
        name: permissionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/Entity.Role'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Detach permission from role
      tags:
      - Role-Permission
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"